[GOB-82] Refactor usando ExcelUtils
